% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-items.R
\name{get_bfi_items}
\alias{get_bfi_items}
\title{Big Five item texts (EN from psych::bfi.dictionary, JA by the first author)}
\usage{
get_bfi_items()
}
\value{
list(en, ja, factor, factor_labels, reversed)
}
\description{
Big Five item texts (EN from psych::bfi.dictionary, JA by the first author)
}
