% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-items.R
\name{panas_ja_validated}
\alias{panas_ja_validated}
\title{Validated Japanese PANAS item wordings (Kawahito et al., 2011)}
\format{
See the source for structure.
}
\usage{
panas_ja_validated
}
\description{
Validated Japanese PANAS item wordings (Kawahito et al., 2011)
}
