% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-items.R
\name{prestige_anchors}
\alias{prestige_anchors}
\title{Pre-specified anchor sets for the occupational-prestige axis}
\format{
See the source for structure.
}
\usage{
prestige_anchors
}
\description{
Pre-specified anchor sets for the occupational-prestige axis
}
