% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-viz.R
\name{save_fig}
\alias{save_fig}
\title{図を指定フォルダに保存（figuresフォルダを自動作成）}
\usage{
save_fig(plot, name, dir = "figures", w = 11, h = 7.5, dpi = 180)
}
\arguments{
\item{plot}{ggplotオブジェクト}

\item{name}{ファイル名（拡張子なし）}

\item{dir}{保存先ディレクトリ}

\item{w, }{h   幅・高さ（インチ）}

\item{dpi}{解像度}
}
\description{
図を指定フォルダに保存（figuresフォルダを自動作成）
}
