% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-items.R
\name{schwartz_items}
\alias{schwartz_items}
\title{Schwartz value descriptions (English) and the theoretical ring order}
\format{
See the source for structure.
}
\usage{
schwartz_items
}
\description{
Schwartz value descriptions (English) and the theoretical ring order
}
