% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-viz.R
\name{trajectory_length}
\alias{trajectory_length}
\title{個人の意味空間上の軌跡を計算}
\usage{
trajectory_length(df, person_col = "person", time_col = "time")
}
\arguments{
\item{df}{pca_2d()$df にperson列とtime列を追加したもの}

\item{person_col}{個人ID列名}

\item{time_col}{時点列名}
}
\value{
個人ごとの軌跡長（total_dist）を含むデータフレーム
}
\description{
個人の意味空間上の軌跡を計算
}
