% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-items.R
\name{valence_anchors}
\alias{valence_anchors}
\title{Pre-specified anchor sets for the valence axis (EN/JA)}
\format{
See the source for structure.
}
\usage{
valence_anchors
}
\description{
Pre-specified anchor sets for the valence axis (EN/JA)
}
