% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-viz.R
\name{within_between_sim}
\alias{within_between_sim}
\title{グループ内・グループ間の平均コサイン類似度と差 Δ}
\usage{
within_between_sim(sim_mat, groups)
}
\arguments{
\item{sim_mat}{cos_sim_matrix() の出力}

\item{groups}{各行に対応するグループラベル (character vector)}
}
\value{
list(within, between, delta = within - between, ratio)
}
\description{
グループ内・グループ間の平均コサイン類似度と差 Δ
}
